Close the 1.0 review: API freeze, CopyFrom fidelity, reader bounds
Reader and writer fixes from the 2026-09-06 review, each reproduced
before it was changed and pinned by a test:

- CopyFrom from a generic fs.FS (fstest.MapFS, embed.FS, os.DirFS off
  linux/darwin) dropped every modification time; add falls back to
  info.ModTime() when Sys() carries none.
- The copyFromImage fast path silently continued where the reader errors
  (missing shared xattr, bad dirent offset, truncated inline dir or
  symlink, inline data crossing its block); a span helper now returns
  exactly the bytes asked for or an error.
- Flat-plain data addresses are bounded by the image size in loadBlock,
  openDirect and buildDataRanges, as chunk addresses already were.
- Hard links are detected across CopyFrom by (Dev, Ino) and become
  aliases; a file's nlink is computed from the names the image holds.
- Ownership and mtime on the BSDs: mkfs_stat.go plus one file per
  Stat_t timestamp spelling replace the linux/darwin pair.
- TestDataRangeSparseChunkBased and ExampleOpen build their images at
  test time instead of referencing a deleted testdata directory.
- Upstream 03d68d8 ported: ACL xattr prefixes without the trailing dot.

API freeze for v1: the deprecated Opt and EroFS are gone; Writer.Mknod
takes an fs.FileMode and accepts only device, FIFO and socket types;
Stat's Ino, Nlink and Rdev are uint64 to match its accessor interfaces
and InodeLayout is no longer exported; DataRange is documented; Chown
rejects ids outside 32 bits; WithDataFile takes a DataFile interface
(Writer, Seeker, ReaderAt, which Writer.Open needs); the concurrency
contract and hard limits are in the package documentation.

+++ internal/erofstest/check.go
@@ -65,7 +65,7 @@ func CheckDevice(t testing.TB, fsys fs.FS, name string, ftype fs.FileMode, rdev 
 	if st.Mode&fs.ModeType != ftype {
 		t.Errorf("%s: type %v, want %v", name, st.Mode&fs.ModeType, ftype)
 	}
-	if st.Rdev != rdev {
+	if st.Rdev != uint64(rdev) {
 		t.Errorf("%s: rdev %d, want %d", name, st.Rdev, rdev)
 	}
 }
